재무 데이터 정밀도 향상을 위한 자료형 변경 및 데이터 복구 API 구현

+++ src/main/java/com/stock/batch/batchJob/CorpFinanceBatch.java
@@ -116,7 +116,7 @@ public class CorpFinanceBatch {
         return builder.build();
     }
 
-    private Double calculateGrowthRate(Integer currentValue, Integer previousValue) {
+    private Double calculateGrowthRate(Long currentValue, Long previousValue) {
         if (currentValue == null || previousValue == null || previousValue == 0) {
             return null;
         }
